% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggregate.swmpr}
\alias{aggregate.swmpr}
\title{Aggregate swmpr data by specified time period and method}
\usage{
\method{aggregate}{swmpr}(swmpr_in, by, FUN = mean, params = NULL,
  na.action = na.pass, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{by}{chr string of time period for aggregation one of \code{'years'}, \code{'quarters'}, \code{'months'}, \code{'weeks'}, \code{'days'}, or \code{'hours'}}

\item{FUN}{aggregation function, default \code{mean}}

\item{params}{names of parameters to aggregate, default all}

\item{na.action}{function for treating missing data, default \code{na.pass}}
}
\value{
Returns an aggregated swmpr object. QAQC columns are removed if included with input object.
}
\description{
Aggregate swmpr data by specified time period and method
}

