% Generated by roxygen2 (4.0.2): do not edit by hand
\name{all_params_dtrng}
\alias{all_params_dtrng}
\title{Get station records from the CDMO for a date range, CDMO equivalent of \code{exportAllParamsDateRangeXMLNew}}
\usage{
all_params_dtrng(station_code, dtrng, param = NULL)
}
\arguments{
\item{station_code}{chr string of station, 7 or 8 characters}

\item{dtrng}{two element chr string, each of format MM/DD/YYYY}

\item{param}{chr string for a single parameter to return, defaults to all parameters for a station type.}
}
\value{
Returns a swmpr object, all parameters for a station type (nutrients, water quality, or meteorological) or a single parameter if specified.  QAQC columns are not provided for single parameters.  Up to 1000 records can be obtained.
}
\description{
Get station records from the CDMO for a date range, CDMO equivalent of \code{exportAllParamsDateRangeXMLNew}
}

