% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmpr_misc.R
\docType{data}
\name{apacpnut}
\alias{apacpnut}
\title{Example nutrient data for Apalachicola Bay Cat Point station.}
\format{A \code{\link{swmpr}} object and \code{\link[base]{data.frame}} with 215 rows and 13 variables:
\describe{
  \item{\code{datetimestamp}}{POSIXct}
  \item{\code{po4f}}{num}
  \item{\code{f_po4f}}{chr}
  \item{\code{nh4f}}{num}
  \item{\code{f_nh4f}}{chr}
  \item{\code{no2f}}{num}
  \item{\code{f_no2f}}{chr}
  \item{\code{no3f}}{num}
  \item{\code{f_no3f}}{chr}
  \item{\code{no23f}}{num}
  \item{\code{f_no23f}}{chr}
  \item{\code{chla_n}}{num}
  \item{\code{f_chla_n}}{chr}
}}
\source{
\url{https://s3.amazonaws.com/swmpexdata/zip_ex.zip}
}
\usage{
apacpnut
}
\description{
An example nutrient dataset for Apalachicola Bay Cat Point station.  The data are a \code{\link{swmpr}} object that have been imported into R from csv files using the \code{\link{import_local}} function.  The raw data were obtained from the CDMO data portal but can also be accessed from a zip file created for this package.  See the source below.  The help file for the \code{\link{import_local}} function describes how the data can be imported from the zip file.  Attributes of the dataset include \code{names}, \code{row.names}, \code{class}, \code{station}, \code{parameters}, \code{qaqc_cols}, \code{date_rng}, \code{timezone}, and \code{stamp_class}.
}
\examples{
data(apacpnut)
}
\keyword{datasets}

