% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swmpr_misc.R
\docType{data}
\name{apaebmet}
\alias{apaebmet}
\title{Example weather data for Apalachicola Bay East Bay station.}
\format{A \code{\link{swmpr}} object and \code{\link[base]{data.frame}} with 70176 rows and 23 variables:
\describe{
  \item{\code{datetimestamp}}{POSIXct}
  \item{\code{atemp}}{num}
  \item{\code{f_atemp}}{chr}
  \item{\code{rh}}{num}
  \item{\code{f_rh}}{chr}
  \item{\code{bp}}{num}
  \item{\code{f_bp}}{chr}
  \item{\code{wspd}}{num}
  \item{\code{f_wspd}}{chr}
  \item{\code{maxwspd}}{num}
  \item{\code{f_maxwspd}}{chr}
  \item{\code{wdir}}{num}
  \item{\code{f_wdir}}{chr}
  \item{\code{sdwdir}}{num}
  \item{\code{f_sdwdir}}{chr}
  \item{\code{totpar}}{num}
  \item{\code{f_totpar}}{chr}
  \item{\code{totprcp}}{num}
  \item{\code{f_totprcp}}{chr}
  \item{\code{cumprcp}}{num}
  \item{\code{f_cumprcp}}{chr}
  \item{\code{totsorad}}{num}
  \item{\code{f_totsorad}}{chr}
}}
\source{
\url{https://s3.amazonaws.com/swmpexdata/zip_ex.zip}
}
\usage{
apaebmet
}
\description{
An example weather dataset for Apalachicola Bay East Bay station.  The data are a \code{\link{swmpr}} object that have been imported into R from csv files using the \code{\link{import_local}} function.  The raw data were obtained from the CDMO data portal but can also be accessed from a zip file created for this package.  See the source below.  The help file for the \code{\link{import_local}} function describes how the data can be imported from the zip file.  Attributes of the dataset include \code{names}, \code{row.names}, \code{class}, \code{station}, \code{parameters}, \code{qaqc_cols}, \code{date_rng}, \code{timezone}, and \code{stamp_class}.
}
\examples{
data(apaebmet)
}
\keyword{datasets}

