% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_modeled_par.R
\name{calc_modeled_par}
\alias{calc_modeled_par}
\title{Calculate modeled photosynthetically active radiation}
\usage{
calc_modeled_par(doy, time_hrs, rh, temp, bp_mb, lat_decdeg, long_decdeg,
  longTZ)
}
\arguments{
\item{doy}{Julian day of year}

\item{time_hrs}{time of day, as whole or decimal hours}

\item{rh}{relative humidity, percent}

\item{temp}{temperature, degrees Celsius}

\item{bp_mb}{barometric pressure, millibars}

\item{lat_decdeg}{latitude, in decimal degrees}

\item{long_decdeg}{longitude, in decimal degrees, but POSITIVE FOR WEST longitudes}

\item{longTZ}{center longitude of the time zone: 75, 90, 105, 120 for eastern, central, mountain, and pacific, respectively. Others can be found here: http://clearskycalculator.com/longitudeTZ.htm}
}
\value{
a single value. units are micromoles/m^2/s  (NOTE that SWMP PAR data values are totals for 15-minute PAR, in mmol/15-min. To convert between units: umol/m^2/s = 1.11 * mmol/15-min)
}
\description{
Based on the clearskycalculator.com quantum sensor model
}
\details{
This function models photosynthetically active radiation (PAR) for a given location, at a given point in time, based on other meteorological parameters that can affect radiation: day of year (which influences sun angle), temperature, relative humidity, and barometric pressure.  

I THINK that I used barometric pressure in lieu of elevation (which is what the website uses). CHECK ON THIS. 

Results are about 1.5% different than those from the clearskycalculator website, presumably due to rounding.
}
\examples{
 
# get location info for apaebmet
apaeb_loc <- stat_locs[stat_locs$station_code == "apaeb", ]

# calculate modeled PAR for noon on Jan. 1st, 2012; 
# actual was 903.1 mmol/15-min; this converts to 1002 umol/m^2/s
calc_modeled_par(doy = 1, time_hrs = 12, 
                 rh = 98, temp = 16.9, bp_mb = 1024,
                 lat_decdeg = apaeb_loc$latitude,
                 long_decdeg = -1 * apaeb_loc$longitude,
                 longTZ = 75)
                 
# calculate for noon on June 1st, 2012; 
# actual was 1843.1 mmol/15-min; this converts to 2045.8 umol/m^2/s  
calc_modeled_par(doy = 152, time_hrs = 12, 
                 rh = 64, temp = 27.8, bp_mb = 1011,
                 lat_decdeg = apaeb_loc$latitude,
                 long_decdeg = -1 * apaeb_loc$longitude,
                 longTZ = 75)
}
\references{
This is based on the model used on clearskycalculator.com.
website: http://clearskycalculator.com/quantumsensor.htm
ASCE report with derivation of model:
https://www.apogeeinstruments.com/content/EWRI-ASCE-Reference-ET.pdf
Appendices (Appendix D has equations used here):
https://www.apogeeinstruments.com/content/EWRI-ASCE-Reference-ET-Appendices.pdf
}
\concept{analyze}
