% Generated by roxygen2 (4.0.2): do not edit by hand
\name{decomp}
\alias{decomp}
\alias{decomp.swmpr}
\title{Seasonal trend decomposition of swmpr data}
\usage{
decomp(swmpr_in, ...)

\method{decomp}{swmpr}(swmpr_in, param, s.window, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{...}{arguments passed to \code{stl} and other methods}

\item{param}{chr string of swmpr parameter to decompose}

\item{s.window}{either the chrr string \code{"periodic"} or the span (in lags) of the loess window for seasonal extraction, which should be odd.}
}
\value{
Returns an stl object.
}
\description{
Decompose swmpr data into seasonal, trend, and irregular components using \code{\link[stats]{stl}} and \code{\link[stats]{loess}}
}
\details{
This function is a simple wrapper to the \code{\link[stats]{stl}} function written by Brian Ripley.  In general, \code{stl} decomposes a time series using \code{\link[stats]{loess}} smoothing into separate seasonal, trend, and remainder components. All arguments for the \code{stl} function are applicable, although each have default values excluding s.window.  All methods available for stl objects also apply to the output, including \code{plot.stl}.
}
\references{
R. B. Cleveland, W. S. Cleveland, J.E. McRae, and I. Terpenning (1990) STL: A Seasonal-Trend Decomposition Procedure Based on Loess. Journal of Official Statistics, 6, 3–73.
}

