% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swmpr_analyze.R
\name{ecometab}
\alias{ecometab}
\alias{ecometab.swmpr}
\title{Ecosystem metabolism}
\usage{
ecometab(swmpr_in, ...)

\method{ecometab}{swmpr}(swmpr_in, depth_val = NULL, met_units = "mmol",
  trace = FALSE, ...)
}
\arguments{
\item{swmpr_in}{Input swmpr object which must include time series of dissolved oxygen,}

\item{...}{arguments passed to other methods}

\item{depth_val}{numeric value for station depth if time series is not available}

\item{met_units}{chr indicating units of output for oxygen, either as mmol or grams}

\item{trace}{logical indicating if progress is shown in the console}
}
\value{
The original \code{\link{swmpr}} object is returned that includes a metabolism attribute as a \code{\link[base]{data.frame}} of daily integrated metabolism estimates.  See the examples for retrieval.
\describe{
 \item{\code{date}}{The metabolic day for each estimate, defined as the approximate 24 hour period between sunsets on adjacent calendar days.}
 \item{\code{ddo}}{Mean hourly rate of change for DO, mmol/m3/hr}
 \item{\code{D}}{Mean air-sea gas exchange of DO, mmol/m2/hr}
 \item{\code{DOF_d}}{Mean DO flux during day hours, mmol/m2/hr}
 \item{\code{D_d}}{Mean air-sea gas exchange of DO during day hours, mmol/m2/hr}
 \item{\code{DOF_n}}{Mean DO flux during night hours, mmol/m2/hr}
 \item{\code{D_n}}{Mean air-sea gas exchange of DO during night hours, mmol/m2/hr}
 \item{\code{Pg}}{Gross production, O2 mmol/m2/d}
 \item{\code{Rt}}{Total respiration, O2 mmol/m2/d}
 \item{\code{NEM}}{Net ecosytem metabolism, O2 mmol/m2/d}
 \item{\code{Pg_vol}}{Volumetric gross production,  O2 mmol/m3/d}
 \item{\code{Rt_vol}}{Volumetric total respiration, O2 mmold/m3/d}
}
}
\description{
Estimate ecosystem metabolism using the Odum open-water method.  Estimates of daily integrated gross production, total respiration, and net ecosystem metabolism are returned.
}
\details{
Input data must be combined water quality and weather datasets.  This is typically done using the \code{\link{comb}} function after creating separate swmpr objects.

The open-water method is a common approach to quantify net ecosystem metabolism using a mass balance equation that describes the change in dissolved oxygen over time from the balance between photosynthetic and respiration rates, corrected for air-sea gas diffusion at the surface. The air-sea gas exchange is estimated as the difference between the DO saturation concentration and observed DO concentration, multiplied by a volumetric reaeration coefficient (see the appendix in Thebault et al. 2008).  The diffusion-corrected DO flux estimates are averaged during day and night for each 24 hour period in the time series, where flux is an hourly rate of DO change. DO flux is averaged during night hours for respiration and averaged during day hours for net production. Respiration rates are assumed constant during day and night such that total daily rates are calculated as hourly respiration multiplied by 24. The metabolic day is considered the approximate 24 hour period between sunsets on two adjacent calendar days.  Respiration is subtracted from daily net production estimates to yield gross production.

Volumetric rates for gross production and total respiration are based on total depth of the water column, which is assumed to be mixed.  Water column depth is based on mean value for the depth variable across the time series in the \code{\link{swmpr}} object and is floored at 1 meter for very shallow stations.  Additionally, the volumetric reaeration coefficient requires an estimate of the anemometer height of the weather station, which is set as 10 meters by default.  The metadata should be consulted for exact height.  The value can be changed manually using a \code{height} argument, which is passed to \code{\link{calcKL}}.

Metabolism estimates are not returned for dates with less than three hourly records during either day or night periods. Missing values for air temperature, barometric pressure, and wind speed are replaced with the climatological means for the period of record.  Climatological means are based on hourly average values by month.

All estimates are in mmol of oxygen but can be converted  to grams by changing the default arguments (i.e., 1mmol O2 = 32 mg O2, 1000 mg = 1g, multiply all estimates by 32/1000).

The specific approach for estimating metabolism is described in more detail in Caffrey et al. 2013.
}
\examples{
## import water quality and weather data
data(apadbwq)
data(apaebmet)

## qaqc, combine
wq <- qaqc(apadbwq)
met <- qaqc(apaebmet)
dat <- comb(wq, met)

## estimate metabolism
res <- ecometab(dat, trace = TRUE)

## change height (m) of weather station anemometer
res <- ecometab(dat, trace = TRUE, height = 5)
res <- attr(res, 'metabolism')

## output units in grams of oxygen
res <- ecometab(dat, trace = TRUE, met_units = 'grams')
res <- attr(res, 'metabolism')
}
\references{
Caffrey JM, Murrell MC, Amacker KS, Harper J, Phipps S, Woodrey M. 2013. Seasonal and inter-annual patterns in primary production, respiration and net ecosystem metabolism in 3 estuaries in the northeast Gulf of Mexico. Estuaries and Coasts. 37(1):222-241.

Odum HT. 1956. Primary production in flowing waters. Limnology and Oceanography. 1(2):102-117.

Thebault J, Schraga TS, Cloern JE, Dunlavey EG. 2008. Primary production and carrying capacity of former salt ponds after reconnection to San Francisco Bay. Wetlands. 28(3):841-851.
}
\seealso{
\code{\link{comb}}, \code{\link{plot_met}}
}

