% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swmpr_analyze.R
\name{hist.swmpr}
\alias{hist.swmpr}
\title{Plot swmpr using a histogram}
\usage{
\method{hist}{swmpr}(x, subset = NULL, select, operator = NULL, ...)
}
\arguments{
\item{x}{input swmpr object}

\item{subset}{chr string of form 'YYYY-mm-dd HH:MM' to subset a date range.  Input can be one (requires \code{operator} or two values (a range), passed to \code{\link{subset}}.}

\item{select}{chr string of parameters to keep, passed to \code{\link{subset}}.}

\item{operator}{chr string specifiying binary operator (e.g., \code{'>'}, \code{'<='}) if subset is one date value, passed to \code{\link{subset}}.}

\item{...}{other arguments passed to \code{\link[graphics]{hist}}}
}
\description{
Plot a histogram showing the distribution of a swmpr parameter
}
\details{
The swmpr method for histograms is a convenience function for the default histogram function.  Conventional histogram methods also work well since swmpr objects are also data frames.  See the examples for use with different methods.
}
\examples{
## get data
path <- system.file('zip_ex', package = 'SWMPr')
dat <- import_local(path, 'apadbwq')

## plot using swmpr method, note default line plot
hist(dat, select = 'do_mgl')

## plot using defualt method
hist(dat$do_mgl)
}
\seealso{
\code{\link[graphics]{hist}}
}

