% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swmpr_analyze.R
\name{met_day}
\alias{met_day}
\title{Identify metabolic days in a swmpr time series}
\usage{
met_day(dat_in, stat_in)
}
\arguments{
\item{dat_in}{data.frame}

\item{stat_in}{chr vector of station name including data type}
}
\description{
Identify metabolic days in a swmpr time series based on sunrise and sunset times for a location and date.  The metabolic day is considered the 24 hour period between sunsets for two adjacent calendar days.
}
\details{
This function is only used within \code{\link{ecometab}} and should not be called explicitly.
}
\seealso{
\code{\link{ecometab}}
}

