% Generated by roxygen2 (4.0.2): do not edit by hand
\name{na.approx.swmpr}
\alias{na.approx.swmpr}
\title{Linearly interpolate gaps in swmpr data}
\usage{
\method{na.approx}{swmpr}(swmpr_in, params = NULL, maxgap, na.rm = F, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{params}{is chr string of swmpr parameters to smooth, default all}

\item{maxgap}{numeric vector indicating maximum gap size to interpolate where size is numer of records, must be explicit}

\item{na.rm}{logical. If the result of the interpolation includes NAs, should these be removed?}
}
\value{
Returns a swmpr object. QAQC columns are removed if included with input object.
}
\description{
Linearly interpolate gaps in swmpr data
}

