% Generated by roxygen2 (4.0.2): do not edit by hand
\name{param_names}
\alias{param_names}
\title{Get parameter column names for each parameter type: nutrients, water quality, or meteorological}
\usage{
param_names(param_type = c("nut", "wq", "met"))
}
\arguments{
\item{param_type}{chr string specifying \code{'nut'}, \code{'wq'}, or \code{'met'}.  Input can be one to three types.}
}
\value{
Returns a named list of parameters for the \code{param_type}.  The parameter names are lower-case strings of SWMP parameters and corresponding qaqc names (\code{'f_'} prefix)
}
\description{
Get parameter column names for each parameter type: nutrients, water quality, or meteorological
}

