% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parser}
\alias{parser}
\title{Parsing function for objects returned from SOAP server}
\usage{
parser(soap_in, parent_in = "data")
}
\arguments{
\item{soap_in}{soap object returned from CDMO server}

\item{parent_in}{chr string of parent nodes to parse}
}
\value{
Returns a \code{data.frame} of parsed XML nodes
}
\description{
Parsing function for objects returned from SOAP server
}

