% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_summary}
\alias{plot_summary}
\alias{plot_summary.swmpr}
\title{Plot graphical summaries of SWMP data}
\usage{
plot_summary(swmpr_in, ...)

\method{plot_summary}{swmpr}(swmpr_in, param, years = NULL, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{...}{additional arguments passed to other methods, currently not used}

\item{param}{chr string of variable to plot}

\item{years}{numeric vector of starting and ending years to plot, default all}
}
\value{
A graphics object (Grob) of multiple \code{\link[ggplot2]{ggplot}} objects.
}
\description{
Plot graphical summaries of SWMP data for individual parameters, including seasonal/annual trends and anomalies
}
\details{
This function creates several graphics showing seasonal and annual trends for a given swmp parameter.  Plots include aggregated monthly distributions, monthly anomalies, and annual anomalies.  Anomalies are defined as the difference between the monthly or annual average from the grand mean.  Monthly anomalies are in relation to the grand mean for the same month across all years.  This function currently only works for water qualiyt and weather data.  An interactive Shiny widget is available: \url{https://beckmw.shinyapps.io/swmp_summary/}
}
\examples{
## import data
path <- system.file('zip_ex', package = 'SWMPr')
dat <- import_local(path, 'apacpwq')
dat <- qaqc(dat, qaqc_keep = NULL)

## plot
plot_summary(dat)
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}

