% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qaqcchk}
\alias{qaqcchk}
\alias{qaqcchk.swmpr}
\title{Summary of QAQC flags in SWMP data}
\usage{
qaqcchk(swmpr_in, ...)

\method{qaqcchk}{swmpr}(swmpr_in)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{...}{arguments passed to other methods}
}
\value{
Returns a \code{\link[base]{data.frame}} with columns for swmpr parameters and row counts indicating the number of observations in each parameter assigned to a flag value.
}
\description{
Summary of the number of observations with a given QAQC flag for each parameter column of a swmpr object
}
\seealso{
qaqc
}

