% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmpr_analyze.R
\name{quant.plot}
\alias{quant.plot}
\title{Generate a plot where a single year's data is overlaid on historical data.}
\usage{
\method{quant.plot}{swmpr}(data, paramtoplot, yr, yrstart, yrend,
  yaxislab = paramtoplot, yrcolor = "red3", bgcolor1 = "lightgray",
  bgcolor2 = "gray65", maintitle = paste0(toupper(attributes(data)$station),
  " ", yr, " Daily Average ", paste0(toupper(substr(paramtoplot, 1, 1)),
  substr(paramtoplot, 2, nchar(paramtoplot))), "\\noverlaid on ", yrstart,
  " - ", yrend, " daily averages"), ...)
}
\arguments{
\item{data}{Input data object.}

\item{paramtoplot}{Parameter of interest. Must be in quotation marks and 
must exactly match the column name of the data frame.}

\item{yr}{Year to feature as a line on the plot.}

\item{yrstart}{Year to begin range of comparison data}

\item{yrend}{Year to end range of comparison data}

\item{yaxislab}{Label for y-axis. Default is name of \code{paramtoplot}.}

\item{yrcolor}{Line color for year of interest. Default is red3.}

\item{bgcolor1}{Color for outer 50\% of data range. Default is lightgray.}

\item{bgcolor2}{Color for middle 50\% of data range. Default is gray65.}

\item{maintitle}{Plot title. Default pastes together site name, parameter name, year to feature, and range of years to use for comparison, e.g. "GNDBHWQ 2017 Daily Average Temp overlaid on 2006-2016 daily averages".}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{ggplot}}}
}
\value{
An R plot created using ggplot2
}
\description{
A line for a single year is plotted over ribbons of historical data quantiles.
}
\details{
This function aggregates data into daily average values for the entire time 
series. Then it generates quantiles (min, 25\%, 75\%, max) for each individual 
calendar day (01/01, 01/02, ... 12/31). These quantiles make up a ribbon plot
of historical data, and a line is plotted on top of this to represent data
from a single year in this historical context.

required packages: dplyr, lubridate, ggplot2, tibble
assuming data originally imported and processed with SWMPr
}
\examples{
# use data from Apalachicola NERR's 
# Cat Point water quality station 
data(apacpwq)
# qaqc the data
apacpwq <- qaqc(apacpwq)

# generate a plot of salinity for 2013 overlaid on 2012-2013 data
quant.plot(apacpwq, 'sal', yr = 2013, yrstart = 2012, yrend = 2013)

# change some of the defaults
quant.plot(apacpwq, 'sal', yr = 2013, yrstart = 2012, yrend = 2013, bgcolor1 = "lightsteelblue2", bgcolor2 = "lightsteelblue4", yaxislab = "Salinity (psu)")

}
\author{
Kim Cressman, Marcus Beck
}
\concept{
analyze
}
