% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setstep}
\alias{setstep}
\title{Create a continuous time vector at set time step for a swmpr object}
\usage{
setstep(x, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{timestep}{numeric value of time step to use in minutes}

\item{differ}{numeric value defining buffer for merging time stamps to standardized time series}
}
\value{
Returns a swmpr object for the specified time step
}
\description{
Create a continuous time vector at set time step for a swmpr object
}

