% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swmpr_misc.R
\docType{data}
\name{stat_locs}
\alias{stat_locs}
\title{Locations of NERRS sites}
\format{A \code{\link[base]{data.frame}} object with 161 rows and 3 variables:
\describe{
  \item{\code{station_code}}{chr}
  \item{\code{latitude}}{numeric}
  \item{\code{longitude}}{numeric}
}}
\usage{
stat_locs
}
\description{
Location of NERRS sites in decimal degress.  Only active sites as of January 2015 are included.  Sites are identified by five letters indicing the reserve and site names.  The dataset is used to plot locations with the \code{\link{map_reserve}} function if the user's computer is not able to access the CDMO online web services.
}
\seealso{
\code{\link{map_reserve}}
}
\keyword{datasets}

