% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.swmpr}
\alias{summary.swmpr}
\title{Basic summary of swmpr data}
\usage{
\method{summary}{swmpr}(swmpr_in, qaqc_summ = F, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{...}{other arguments passed to \code{summary}}
}
\description{
Basic summary of swmpr data for numeric, factor, and character strings, see the documentation for \code{\link[base]{summary}}
}

