% Generated by roxygen2 (4.0.2): do not edit by hand
\name{swmpr}
\alias{swmpr}
\title{Wrapper for creating a swmpr object using \code{structure}}
\usage{
swmpr(stat_in, meta_in)
}
\arguments{
\item{stat_in}{\code{data.frame} of swmp data}

\item{meta_in}{chr string for station code (7 or 8 characters), can be multiple stations if data are combined}
}
\value{
Returns a swmpr object to be used with S3 methods
}
\description{
Wrapper for creating a swmpr object using \code{structure}
}

