% Generated by roxygen2 (4.0.2): do not edit by hand
\name{time_vec}
\alias{time_vec}
\title{Convert \code{datetimestamp} string to \code{POSIXct}, using correct timezone for a station (no daylight savings)}
\usage{
time_vec(chr_in = NULL, station_code, tz_only = F)
}
\arguments{
\item{chr_in}{chr string of datetimestamp vector}

\item{station_code}{is chr string for station (three or more characters)}

\item{tz_only}{logical that returns only the timezone, default \code{F}}
}
\value{
Returns a POSIX vector if \code{tz_only} is true, otherwise the timezone for a station is returned as a chr string
}
\description{
Convert \code{datetimestamp} string to \code{POSIXct}, using correct timezone for a station (no daylight savings)
}

