% Generated by roxygen2 (4.0.2): do not edit by hand
\name{all_params}
\alias{all_params}
\title{Import current station records from the CDMO}
\usage{
all_params(station_code, Max = 100)
}
\arguments{
\item{station_code}{chr string of station, 7 or 8 characters}

\item{Max}{numeric value for number of records to obtain from the current date, maximum of 100}
}
\value{
Returns a swmpr object, all available parameters including QAQC columns
}
\description{
Import current station records from the CDMO starting with the most current date, CDMO equivalent of \code{exportAllParamsXMLNew}
}

