% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comb}
\alias{comb}
\alias{comb.swmpr}
\title{Combine swmpr data}
\usage{
comb(...)

\method{comb}{swmpr}(..., timestep = 15, differ = timestep/2,
  method = "union")
}
\arguments{
\item{...}{swmpr object input from one to many}

\item{timestep}{numeric value of time step to use in minutes, passed to \code{setstep}}

\item{differ}{numeric value defining buffer for merging time stamps to standardized time series, passed to \code{setstep}}

\item{method}{chr string indicating method of combining (\code{'union'} for all dates as continuous time series, \code{'intersect'} for areas of overlap, or \code{'station'} for a given station)}
}
\value{
Returns a combined swmpr object
}
\description{
Combine swmpr data types for a station by common time series
}
\details{
The \code{comb} function is used to combine multiple swmpr objects into a single object with a continuous time series at a given step. The \code{timestep} function is used internally such that \code{timestep} and \code{differ} are accepted arguments for \code{comb}.

The function requires one or more swmpr objects as input as separate, undefined arguments. The remaining arguments must be called explicitly since an arbitrary number of objects can be used as input. In general, the function combines data by creating a master time series that is used to iteratively merge all swmpr objects. The time series for merging depends on the value passed to the \code{method} argument. Passing \code{'union'} to \code{method} will create a time series that is continuous starting from the earliest date and the latest date for all input objects. Passing \code{'intersect'} to \code{method} will create a time series that is continuous from the set of dates that are shared between all input objects. Finally, a seven or eight character station name passed to \code{method} will merge all input objects based on a continuous time series for the given station. The specified station must be present in the input data. Currently, combining data types from different stations is not possible, excluding weather data which are typically at a single, dedicated station.
}
\examples{
## get nuts, wq, and met data as separate objects for the same station
## note that most sites usually have one weather station
path <- system.file('zip_ex', package = 'SWMPr')
swmp1 <- import_local(path, 'apacpnut')
swmp2 <- import_local(path, 'apacpwq')
swmp3 <- import_local(path, 'apaebmet')

## combine nuts and wq data by union
comb(swmp1, swmp2, method = 'union')

## combine nuts and wq data by intersect
comb(swmp1, swmp3, method = 'intersect')

## combine nuts, wq, and met data by nuts time series, two hour time step
comb(swmp1, swmp2, swmp3, timestep = 120, method = 'apacpnut')
}
\seealso{
\code{\link{setstep}}
}

