% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comb}
\alias{comb}
\title{Combine swmpr data types for a station by common time series}
\usage{
comb(...)
}
\arguments{
\item{timestep}{numeric value of time step to use in minutes, passed to \code{setstep}}

\item{differ}{numeric value defining buffer for merging time stamps to standardized time series, passed to \code{setstep}}

\item{method}{chr string indicating method of combining (\code{'union'} for all dates as continuous time series, \code{'intersect'} for areas of overlap, or \code{'station'} for a given station)}
}
\value{
Returns a combined swmpr object
}
\description{
Combine swmpr data types for a station by common time series
}

