% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hist.swmpr}
\alias{hist.swmpr}
\title{Plot swmpr using a histogram}
\usage{
\method{hist}{swmpr}(swmpr_in, subset = NULL, select, operator = NULL, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{subset}{chr string of form 'YYYY-mm-dd HH:MM' to subset a date range.  Input can be one (requires \code{operator} or two values (a range), passed to \code{\link{subset}}.}

\item{select}{chr string of parameters to keep, passed to \code{\link{subset}}.}

\item{operator}{chr string specifiying binary operator (e.g., \code{'>'}, \code{'<='}) if subset is one date value, passed to \code{\link{subset}}.}

\item{...}{other arguments passed to \code{\link[graphics]{histogram}}}
}
\description{
Plot a histogram showing the distribution of a swmpr parameter
}

