% Generated by roxygen2 (4.0.2): do not edit by hand
\name{import_local}
\alias{import_local}
\title{Import local CDMO data}
\usage{
import_local(path, station_code, trace = FALSE)
}
\arguments{
\item{path}{chr string of full path to .csv files with raw data}

\item{station_code}{chr string of station to import, typically 7 or 8 characters but may include full name with year, excluding file extension}

\item{trace}{logical indicating if progress is sent to console, default \code{FALSE}}
}
\value{
Returns a swmpr object with all parameters and QAQC columns for the station.  The full date range in the raw data are also imported.
}
\description{
Import local data that were obtained from the CDMO through the zip downloads feature
}
\details{
The function is designed to import local data that were downloaded from the CDMO outside of R. This approach works best for larger data requests.  The function is designed for data from the zip downloads feature in the advanced query section of the CDMO. The function may also work using data from the data export system, but this feature has not been extensively tested (expect bugs). The downloaded data will be in a compressed folder that includes multiple .csv files by year for a given data type (e.g., apacpwq2002.csv, apacpwq2003.csv, apacpnut2002.csv, etc.). The import_local function can be used after the folder is decompressed.

Zip download request through CDMO: \url{http://cdmo.baruch.sc.edu/aqs/zips.cfm}
}
\examples{
## this is the path for csv example files
path <- system.file('zip_ex', package = 'SWMPr')

## import, do not include file extension
import_local(path, 'apaebmet')
}
\seealso{
\code{\link{all_params}}, \code{\link{all_params_dtrng}}, \code{\link{single_param}}
}

