% Generated by roxygen2 (4.0.2): do not edit by hand
\name{import_local}
\alias{import_local}
\title{Import local data that were obtained from the CDMO through a zip request}
\usage{
import_local(path, station_code, trace = F)
}
\arguments{
\item{path}{chr string of full path to .csv files with raw data}

\item{station_code}{chr string of station to import, 7 or 8 characters}

\item{trace}{logical indicating if progress is sent to console, default \code{F}}
}
\value{
Returns a swmpr object with all parameters and QAQC columns for the station.  The full date range in the raw data are also imported.
}
\description{
Import local data that were obtained from the CDMO through a zip request
}

