% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swmpr_misc.R
\name{map_reserve}
\alias{map_reserve}
\title{Map a reserve}
\usage{
map_reserve(nerr_site_id, zoom = 11, text_sz = 6, text_col = "black",
  map_type = "terrain")
}
\arguments{
\item{nerr_site_id}{chr string of the reserve to map, first three characters used by NERRS}

\item{zoom}{numeric value for map zoom, passed to \code{\link[ggmap]{get_map}}}

\item{text_sz}{numeric value for text size of station names, passed to \code{\link[ggplot2]{geom_text}}}

\item{text_col}{chr string for text color of station names, passed to \code{\link[ggplot2]{geom_text}}}

\item{map_type}{chr string indicating the type of base map obtained from Google maps, values are \code{terrain} (default), \code{satellite}, \code{roadmap}, or \code{hybrid}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object for plotting.
}
\description{
Create a map of all the stations in a reserve
}
\details{
This function is a simple wrapper to functions in the ggmap package which returns a map of all of the stations at a NERRS reserve.  The \code{zoom} argument may have to be chosen through trial and error depending on the spatial extent of the reserve.  Additionally, station locations are returned using the \code{site_codes_ind} function if the computer making the request has the IP address registered with CDMO.  Otherwise, a local and possibly outdated file is used.  Instructions for registering with CDMO are online: \url{http://cdmo.baruch.sc.edu/webservices.cfm}
}
\examples{
## defaults

map_reserve('jac')

## satellite map

map_reserve('jac', map_type = 'satellite')

## roadmap

map_reserve('jac', map_type = 'roadmap')

## hybrid

map_reserve('jac', map_type = 'hybrid')
}
\seealso{
\code{\link[ggmap]{get_map}}, \code{\link[ggmap]{ggmap}}, \code{\link[ggplot2]{ggplot}}, \code{site_codes_ind}
}

