% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_names.R
\name{param_names}
\alias{param_names}
\title{Get parameters of a given type}
\usage{
param_names(param_type = c("nut", "wq", "met"))
}
\arguments{
\item{param_type}{chr string specifying \code{'nut'}, \code{'wq'}, or \code{'met'}.  Input can be one to three types.}
}
\value{
Returns a named list of parameters for the \code{param_type}.  The parameter names are lower-case strings of SWMP parameters and corresponding qaqc names (\code{'f_'} prefix)
}
\description{
Get parameter column names for each parameter type
}
\details{
This function is used internally within several functions to return a list of the expected parameters for a given parameter type: nutrients, water quality, or meteorological.  It does not need to be called explicitly.
}
