% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parser}
\alias{parser}
\title{Parse SOAP objects for swmpr}
\usage{
parser(soap_in, parent_in = "data")
}
\arguments{
\item{soap_in}{soap object returned from CDMO server}

\item{parent_in}{chr string of parent nodes to parse}
}
\value{
Returns a \code{data.frame} of parsed XML nodes
}
\description{
Parsing function for objects returned from SOAP server
}
\details{
This function parses XML objects returned from the CDMO SOAP server, which are further passed to \code{\link{swmpr}}.  It is used internally by the data retrieval functions, excluding \code{\link{import_local}}.  The function does not need to be called explicitly.
}

