% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parser}
\alias{parser}
\title{Parse web results for swmpr}
\usage{
parser(resp_in, parent_in = "data")
}
\arguments{
\item{resp_in}{web object returned from CDMO server, response class from httr package}

\item{parent_in}{chr string of parent nodes to parse}
}
\value{
Returns a \code{data.frame} of parsed XML nodes
}
\description{
Parsing function for objects returned from CDMO web services
}
\details{
This function parses XML objects returned from the CDMO server, which are further passed to \code{\link{swmpr}}.  It is used internally by the data retrieval functions, excluding \code{\link{import_local}}.  The function does not need to be called explicitly.
}
