% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.swmpr}
\alias{lines.swmpr}
\alias{plot.swmpr}
\title{Plot swmpr data}
\usage{
\method{plot}{swmpr}(swmpr_in, type = "l", ...)

\method{lines}{swmpr}(swmpr_in, subset = NULL, select, operator = NULL, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{type}{chr string for type of plot, default \code{'l'}.  See \code{\link[graphics]{plot}}.}

\item{...}{other arguments passed to \code{par}, \code{plot.default}, \code{lines}, \code{points}}
}
\description{
Plot a time series of parameters in a swmpr object
}
\details{
The swmpr method for plotting is a convenience function for plotting a univariate time series.  Conventional plotting methods also work well since swmpr objects are also data frames.  See the examples for use with different methods.
}
\examples{
## get data
path <- system.file('zip_ex', package = 'SWMPr')
swmp1 <- import_local(path, 'apadbwq')

## subset
dat <- subset(swmp1, select = 'do_mgl', subset = c('2013-07-01 00:00', '2013-07-31 00:00'))

## plot using swmpr method, note default line plot
plot(dat)

## plot using formula method
plot(do_mgl ~ datetimestamp, dat)

## plot using defualt, add lines
plot(dat$datetimestamp, dat$do_mgl)
lines(dat, select = 'do_mgl', col = 'red')
}
\seealso{
\code{\link[graphics]{plot}}
}

