% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.swmpr}
\alias{lines.swmpr}
\alias{plot.swmpr}
\title{Plot swmpr data}
\usage{
\method{plot}{swmpr}(swmpr_in, type = "l", ...)

\method{lines}{swmpr}(swmpr_in, subset = NULL, select, operator = NULL, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{type}{chr string for type of plot, default \code{'l'}.  See \code{\link[graphics]{plot}}.}

\item{...}{other arguments passed to \code{par}, \code{plot.default}, \code{lines}, \code{points}}
}
\description{
Plot a time series of parameters in a swmpr object
}

