% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qaqc}
\alias{qaqc}
\title{QAQC filtering for data obtained from retrieval functions, local and remote}
\usage{
qaqc(x, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{qaqc_keep}{numeric vector of qaqc flags to keep, default \code{0}}

\item{trace}{logical for progress output on console, default \code{F}}
}
\value{
Returns a swmpr object with \code{NA} values for records that did not match \code{qaqc_keep}.  QAQC columns are also removed.
}
\description{
QAQC filtering for data obtained from retrieval functions, local and remote
}

