% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setstep}
\alias{setstep}
\alias{setstep.swmpr}
\title{Format a swmpr time vctor}
\usage{
setstep(swmpr_in, ...)

\method{setstep}{swmpr}(swmpr_in, timestep = 15, differ = timestep/2)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{...}{arguments passed to other methods}

\item{timestep}{numeric value of time step to use in minutes}

\item{differ}{numeric value defining buffer for merging time stamps to standardized time series}
}
\value{
Returns a swmpr object for the specified time step
}
\description{
Create a continuous time vector at set time step for a swmpr object
}

