% Generated by roxygen2 (4.0.2): do not edit by hand
\name{site_codes}
\alias{site_codes}
\title{Obtain a \code{data.frame} of metadata for all SWMP stations, CDMO equivalent of \code{exportStationCodesXML}}
\usage{
site_codes()
}
\value{
A \code{data.frame} of SWMP metadata
}
\description{
Obtain a \code{data.frame} of metadata for all SWMP stations, CDMO equivalent of \code{exportStationCodesXML}
}

