% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_codes.R
\name{site_codes}
\alias{site_codes}
\title{Obtain metadata for all stations}
\usage{
site_codes()
}
\value{
A \code{data.frame} of SWMP metadata
}
\description{
Obtain a \code{\link[base]{data.frame}} of metadata for all SWMP stations.
}
\details{
This function retrieves data from the CDMO web services.  The computer making the request must have a registered IP address.  Visit the CDMO web services page for more information: \url{http://cdmo.baruch.sc.edu/webservices.cfm}. This is the CDMO equivalent of \code{exportStationCodesXML}.
}
\examples{
\dontrun{

## retrieve metadata for all sites
site_codes()

}
}
\concept{retrieve}
