% Generated by roxygen2 (4.0.2): do not edit by hand
\name{site_codes_ind}
\alias{site_codes_ind}
\title{Get metadata for single SWMP site, CDMO equivalent of \code{NERRFilterStationCodesXMLNew}}
\usage{
site_codes_ind(nerr_site_id)
}
\arguments{
\item{nerr_site_id}{chr string of site, three letters}
}
\value{
An abbreviated \code{data.frame} of the SWMP metadata for the requested site
}
\description{
Get metadata for single SWMP site, CDMO equivalent of \code{NERRFilterStationCodesXMLNew}
}

