% Generated by roxygen2 (4.0.2): do not edit by hand
\name{smoother}
\alias{smoother}
\title{Smooth swmpr data with a moving window average}
\usage{
smoother(x, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{window}{numeric vector of ones defining size of smoothing window, passed to \code{filter}}

\item{sides}{numeric vector defining method of averaging, passed to \code{filter}}

\item{params}{is chr string of swmpr parameters to smooth, default all}
}
\value{
Returns a filtered swmpr object. QAQC columns are removed if included with input object.
}
\description{
Smooth swmpr data with a moving window average
}

