% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmpr_misc.R
\docType{data}
\name{stat_locs}
\alias{stat_locs}
\title{Locations of NERRS sites}
\format{
A \code{\link[base]{data.frame}} object with 161 rows and 4 variables:
\describe{
  \item{\code{station_code}}{chr}
  \item{\code{latitude}}{numeric}
  \item{\code{longitude}}{numeric}
  \item{\code{gmt_off}}{int}
}
}
\usage{
stat_locs
}
\description{
Location of NERRS sites in decimal degrees and time offset from Greenwich mean time.  Only active sites are included.  Sites are identified by five letters indexing the reserve and site names.  The dataset is used to plot locations with the \code{\link{map_reserve}} function and to identify metabolic days with the \code{\link{ecometab}} function.  Created from sampling_locations.csv provided by CDMO.
}
\seealso{
\code{\link{ecometab}}, \code{\link{map_reserve}}
}
\keyword{datasets}
