% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subset.swmpr}
\alias{subset.swmpr}
\title{Subset a swmpr object by a date range, parameters, or non-empty values}
\usage{
\method{subset}{swmpr}(swmpr_in, subset = NULL, select = NULL,
  operator = NULL, rem_rows = F, rem_cols = F)
}
\arguments{
\item{subset}{is chr string of form 'YYYY-mm-dd HH:MM'}

\item{select}{chr string of parameters to keep}

\item{operator}{chr string specifiying binary operator (e.g., \code{'>'}, \code{'<='}) if subset is one date value}

\item{rem_rows}{logical indicating if rows with no data are removed, default \code{F}}

\item{rem_cols}{is logical indicating if cols with no data are removed, default \code{F}}
}
\value{
Returns a swmpr object as a subset of the input.  The original object will be returned if no arguments are specified.
}
\description{
Subset a swmpr object by a date range, parameters, or non-empty values
}

