% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swmpr.R
\name{swmpr}
\alias{swmpr}
\title{Create a swmpr object}
\usage{
swmpr(stat_in, meta_in)
}
\arguments{
\item{stat_in}{\code{data.frame} of swmp data}

\item{meta_in}{chr string for station code (7 or 8 characters), can be multiple stations if data are combined}
}
\value{
Returns a swmpr object to be used with S3 methods
}
\description{
Wrapper for creating a swmpr object
}
\details{
This function is a simple wrapper to \code{\link[base]{structure}} that is used internally within other functions to create a swmpr object.  The function does not have to be used explicitly.  Attributes of a swmpr object include \code{names}, \code{row.names}, \code{class}, \code{station}, \code{parameters}, \code{qaqc_cols}, \code{cens_cols}, \code{date_rng}, \code{timezone}, \code{stamp_class}, \code{metabolism} (if present), and \code{metab_units} (if present).
}
